operation/backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
backports.tarfile==1.2.0
bcrypt==4.0.1
boto3==1.35.95
botocore==1.35.99
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
mangum==0.19.0
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dulwich==0.22.8
ecdsa==0.19.1
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
keyring==25.6.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.0
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pbs-installer==2025.6.4
pkginfo==1.12.1.2
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
xattr==1.1.4
zipp==3.22.0
zstandard==0.23.0
